Multi-arm CMR from a variance rectangle

Compute a shared-control multi-arm CMR allocation from a supplied variance rectangle.

Usage

cmr_multiarm_from_rectangle(rectangle, control = list(), max_vertices = 65536L)

Arguments

rectangle

Multi-arm variance rectangle, either a matrix/data frame with lower and upper columns or a named vector with entries like v_l0, v_u0, v_l1, v_u1.

control

Optional list of solver controls for the general vertex epigraph solver.

max_vertices

Maximum number of hyperrectangle vertices to enumerate.

Value

A list of class cmr_multiarm with named assignment shares pi, regret certificate U_CMR, checked rectangle, enumerated vertices, vertex regrets, binding vertices, and solver diagnostics. For collapsed or full rectangles, closed-form shortcuts are used.

See also

Examples

rect <- c(
  v_l0 = 0.02, v_u0 = 0.08,
  v_l1 = 0.04, v_u1 = 0.12,
  v_l2 = 0.01, v_u2 = 0.07
)
cmr_multiarm_from_rectangle(rect)
#> <cmr_multiarm>
#>   pi: 0=0.398867, 1=0.353341, 2=0.247792
#>   U_CMR: 0.0646975